+++ modules/wasmoon_jit/jit_ffi/jit_internal.h
@@ -123,7 +123,7 @@ int64_t memory_len_desc_internal(wasmoon_memory_t *mem);
 uint8_t *memory_base_desc_internal(wasmoon_memory_t *mem);
 
 // Table operations
-int32_t table_grow_ctx_internal(jit_context_t *ctx, int32_t table_idx, int64_t delta, int64_t init_value);
+int64_t table_grow_ctx_internal(jit_context_t *ctx, int32_t table_idx, int64_t delta, int64_t init_value);
 
 // GC heap management
 void ctx_set_gc_heap_internal(jit_context_t *ctx, GcHeap *heap);
